+++ core-ng/src/main/java/core/framework/impl/queue/RabbitMQ.java
@@ -8,5 +8,6 @@ import com.rabbitmq.client.AMQP;
 public interface RabbitMQ {
     void publish(String exchange, String routingKey, byte[] message, AMQP.BasicProperties properties);
 
+    // consumer must be used in current thread, since the impl uses LockSupport.park()
     RabbitMQConsumer consumer(String queue, int prefetchCount);
 }