+++ core-ng/src/main/java/core/framework/impl/web/response/ByteArrayBody.java
@@ -1,12 +1,16 @@
 package core.framework.impl.web.response;
 
+import core.framework.api.http.ContentType;
+
 /**
  * @author rainbow.cai
  */
 public class ByteArrayBody implements Body {
     final byte[] bytes;
+    final ContentType contentType;
 
-    public ByteArrayBody(byte[] bytes) {
+    public ByteArrayBody(byte[] bytes, ContentType contentType) {
         this.bytes = bytes;
+        this.contentType = contentType;
     }
 }